requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.5
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
uWSGI==2.0.30
Werkzeug==3.1.3
